Move webhooks back up to main instrumentor level
